pyproject.toml
[project]
name = "langgraph-agent"
version = "0.1.0"
description = "langgraph Gemini and Openai agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioconsole>=0.8.1",
    "fastembed>=0.6.1",
    "google-analytics-admin>=0.24.0",
    "google-analytics-data>=0.18.18",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.0.10",
    "langchain-mcp-adapters>=0.0.9",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.1",
    "mcp-use>=1.2.8",
    "mcp[cli]>=1.6.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
]
